Rewrite zk benchmark with criterion
